AMA-1092 Refactor into library and add to blue-ci
